Bugfix: ResolvePath fails when the file name contains [
A file with [  in the name caused the script to abort with a null reference error. Resolve-Path assumes its first argument can contain wildcards and special characters that should be interpreted, causing Resolve-Path to return null in most cases when the file name contains them.  Adding -LiteralPath prevents wildcard expansion and allows [filename].xml to be archived.