Enables Windows Explorer preview pane for a large set of file types
This script will add a perceived type and content type for a large set of file name extensions. 
It adds any extensions that are not already present in Classes Root Registry key.
Any existing perceived type or content type will not be changed. 
It also changes the default value of .nupkg file types to "CompressedFolder" to enable navigation into the package.