Fix for UnauthorizedAccessException on folders
If the $FileSpecs had a spec that matched a folder, an UnauthorizedAccessException would be thrown on CreateEntryFromFile by design: https://msdn.microsoft.com/en-us/library/hh485720(v=vs.110).aspx
Changed filter on Get-ChildItem to exclude containers. 
Note that empty folders can't be created in a zip regardless of this change.