Fix for $filter not being invoked - always true
The curly brackets around $filter were causing it to be treated as a non-invoked scriptblock variable, which evaluates to $true for the purposes of Where-Object. Consequently all files were being archived regardless of timestamp. 
Removed "$_.PSIsContainer -and ..." from Where-Object and added -File switch to Get-ChildItem, to preserve the behavior of skipping directories.