Decodes a JWT AccessToken and converts to Powershell object
JWT (JSON Web Token) access token is a Base64 string with a header and payload delimited by "." but not padded to valid Base64 string length which is divisible by 4.  This script splits and pads the base64 strings, then decodes to ASCII and converts from Json to Powershell objects. 
Unix timestamps in the token are converted to local time.