Optimizations and new features
Tested Get-ID3v1 compared to other methods: BinaryReader, Get-Content, and an Array-only version without any intermediate string:  $enc.GetString($buf[3..30])
This Get-Id3v1 was consistently faster compared to all but the array-only version, and usually faster than it. 

Set-ID3v1 now handles files without an ID3v1 (tested on a file with only ID3v2.3) and with [] in the path. Switched to using "`0" instead of "[Not Specified]" for conciseness.  An empty string can be specified to clear an attribute.